Allow configuring a custom endpoint
The ability to configure a custom URL endpoint for the Okta API is added
to the provider. This is useful for unit testing or running local
caching proxies.

A new configuration option called "custom_endpoint" is added, which by
default takes the value of the "OKTA_CUSTOM_ENDPOINT" environment
variable. When used, and if the provided endpoint is an HTTP endpoint
(rather than an HTTPS endpoint), the HTTPS check is automatically
disabled.

A unit test called `TestCustomEndpoint` is added which also showcases
unit testing of the Okta API without actually contacting Okta.
Information about the new configuration option is also added to the
documentation.