Fix makefile
The makefile of the project would both build and install the provider
binary in the `build` task, which doesn't make much sense. It would also
build it with dynamic linking. Instead, the `build` task will now
statically build the binary, and the previous behavior is moved to the
`install` task.

Also, the makefile allowed setting a different `gofmt` command than the
default (which was `gofumpt`), but `gofmumpt` was still hardcoded in one
of the tasks. This fixes the bug and also sets the default as
`goimports`.